apps/agents/pyproject.toml
[project]
name = "pact-agents"
version = "0.1.0"
description = "Pact AI agent pipeline for contributor rewards"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.1.0",
    "httpx>=0.27.0",
    "apscheduler>=3.10.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
]

[tool.setuptools.packages.find]
include = ["agents*", "routers*", "scheduler*", "models*"]

